--- src/ehrdata/pl/basic.py
@@ -1,63 +1,0 @@
-from anndata import AnnData
-
-
-def basic_plot(adata: AnnData) -> int:
-    """Generate a basic plot for an AnnData object.
-
-    Parameters
-    ----------
-    adata
-        The AnnData object to preprocess.
-
-    Returns
-    -------
-    Some integer value.
-    """
-    print("Import matplotlib and implement a plotting function here.")
-    return 0
-
-
-class BasicClass:
-    """A basic class.
-
-    Parameters
-    ----------
-    adata
-        The AnnData object to preprocess.
-    """
-
-    my_attribute: str = "Some attribute."
-    my_other_attribute: int = 0
-
-    def __init__(self, adata: AnnData):
-        print("Implement a class here.")
-
-    def my_method(self, param: int) -> int:
-        """A basic method.
-
-        Parameters
-        ----------
-        param
-            A parameter.
-
-        Returns
-        -------
-        Some integer value.
-        """
-        print("Implement a method here.")
-        return 0
-
-    def my_other_method(self, param: str) -> str:
-        """Another basic method.
-
-        Parameters
-        ----------
-        param
-            A parameter.
-
-        Returns
-        -------
-        Some integer value.
-        """
-        print("Implement a method here.")
-        return ""
