--- src/ehrdata/pp/basic.py
@@ -1,17 +1,0 @@
-from anndata import AnnData
-
-
-def basic_preproc(adata: AnnData) -> int:
-    """Run a basic preprocessing on the AnnData object.
-
-    Parameters
-    ----------
-    adata
-        The AnnData object to preprocess.
-
-    Returns
-    -------
-    Some integer value.
-    """
-    print("Implement a preprocessing function here.")
-    return 0
