--- src/ehrdata/tl/__init__.py
@@ -1,2 +1,0 @@
-from .basic import basic_tool
-from .omop import get_concept_name
