--- src/ehrdata/tl/basic.py
@@ -1,17 +1,0 @@
-from anndata import AnnData
-
-
-def basic_tool(adata: AnnData) -> int:
-    """Run a tool on the AnnData object.
-
-    Parameters
-    ----------
-    adata
-        The AnnData object to preprocess.
-
-    Returns
-    -------
-    Some integer value.
-    """
-    print("Implement a tool to run on the AnnData object.")
-    return 0
